Feature/dev

+++ StoreMaster/src/main/java/com/akshay/StoreMaster/entity/CartItem.java
@@ -19,9 +19,9 @@ public class CartItem {
     @JoinColumn(name = "product_id", nullable = false)
     private Product product;
 
-    private Integer quantity;
+    private int quantity;
 
-    private int price;
+    private BigDecimal price;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "cart_id", nullable = false)
